feat: Add @radix-ui/react-tabs dependency and update Home page
The code changes add the @radix-ui/react-tabs dependency to the package.json file. This dependency is required for the implementation of the Tabs component in the UI.

In the Home.jsx file, the HotCollections component is added to the Home page. This enhances the content of the page by displaying hot collections.

The index.css file is updated to include styling for a new background color class (.bg_table) and table styling.

The assets/index.js file is modified to export a new image (collect.png) and remove an unused image (nft1.png).

The NewsTrending.jsx component is updated to wrap the content in a section element for better semantic structure.

A new file, tabs.jsx, is added to the components/ui directory. This file contains the implementation of the Tabs component using the @radix-ui/react-tabs library.

These changes improve the functionality and user experience of the application.