refactor: Add new images and update footer component
The code changes in the assets/index.js file add new images (discord.png, twitter.png, insta.png, yt.png) and export them for use in the application.

In the Footer.jsx component, the code is updated to include the new images in the social media links section. The images are displayed as icons for Discord, Twitter, Instagram, and YouTube.

These changes enhance the visual appeal and functionality of the footer component.